+++ CHANGELOG.md
@@ -6,6 +6,37 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [0.5.3] - 2025-12-04
+
+### Changed
+- simplify version retrieval logic in __init__.py
+
+### Other
+- Merge branch 'release/0.5' of https://github.com/alexsc6955/mini-arcade-core into release/0.5
+
+## [0.5.2] - 2025-12-04
+
+### Changed
+- enhance logging in get_version function and add logging configuration
+
+### Other
+- Merge branch 'release/0.5' of https://github.com/alexsc6955/mini-arcade-core into release/0.5
+
+## [0.5.1] - 2025-12-04
+
+### Fixed
+- add print statement to indicate package not found in get_version function
+
+## [0.5.0] - 2025-12-03
+
+### Added
+
+- add version retrieval function and handle exceptions gracefully
+
+### Fixed
+
+- improve docstring for get_version function to clarify return type and exceptions
+
 ## [0.4.0] - 2025-12-03
 
 ### Added